--- close.go
@@ -1,6 +1,0 @@
-package storm
-
-// Close the database
-func (s *DB) Close() error {
-	return s.Bolt.Close()
-}
