--- codec.go
@@ -1,6 +1,0 @@
-package storm
-
-import "github.com/asdine/storm/codec/gob"
-
-// Defaults to gob
-var defaultCodec = gob.Codec
