+++ storm_test.go
@@ -4,9 +4,12 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"reflect"
 	"testing"
 	"time"
 
+	"github.com/asdine/storm/codec/gob"
+	"github.com/asdine/storm/codec/json"
 	"github.com/boltdb/bolt"
 	"github.com/stretchr/testify/assert"
 )
@@ -82,4 +85,48 @@ func (c dummyCodec) Encode(v interface{}) ([]byte, error) {
 
 func (c dummyCodec) Decode(b []byte, v interface{}) error {
 	return nil
+}
+
+func TestCodec(t *testing.T) {
+	u1 := &SimpleUser{Name: "John"}
+	encoded, err := defaultCodec.Encode(u1)
+	assert.Nil(t, err)
+	u2 := &SimpleUser{}
+	err = defaultCodec.Decode(encoded, u2)
+	assert.Nil(t, err)
+	if !reflect.DeepEqual(u1, u2) {
+		t.Fatal("Codec mismatch")
+	}
+}
+
+func TestToBytes(t *testing.T) {
+	b, err := toBytes([]byte("a slice of bytes"), gob.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, []byte("a slice of bytes"), b)
+
+	b, err = toBytes("a string", gob.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, []byte("a string"), b)
+
+	b, err = toBytes(5, gob.Codec)
+	assert.NoError(t, err)
+	assert.NotNil(t, b)
+
+	b, err = toBytes([]byte("Hey"), gob.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, []byte("Hey"), b)
+}
+
+func TestToBytesWithCodec(t *testing.T) {
+	b, err := toBytes([]byte("a slice of bytes"), json.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, []byte("a slice of bytes"), b)
+
+	b, err = toBytes("a string", json.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, []byte("a string"), b)
+
+	b, err = toBytes(&SimpleUser{ID: 10, Name: "John", age: 100}, json.Codec)
+	assert.NoError(t, err)
+	assert.Equal(t, `{"ID":10,"Name":"John"}`, string(b))
 }
